+++ src/robot_SCOUT.hh
@@ -17,7 +17,7 @@ class RobotSCOUT:public Robot
         virtual void SetRGBLED(int channel, uint8_t tl=LED_BLUE, uint8_t tr=0, uint8_t bl=0, uint8_t br=0); 
         virtual bool SetDockingMotor(int channel, int status);
         virtual bool SetHingeMotor(int status);
-        virtual void SetSpeed(int8_t leftspeed, int8_t rightspeed, int8_t sidespeed);
+        virtual void SetSpeed(int leftspeed, int rightspeed, int sidespeed);
         virtual void Reset();
 
         virtual void UpdateSensors();
